created new app: subscriptions; divided models between recipeApp, userAccount, and subscriptions accordingly; added subscription management page
