+++ mysite/settings.py
@@ -41,6 +41,8 @@ INSTALLED_APPS = [
     
     'userAccount.apps.UseraccountConfig',
     'recipeApp.apps.RecipeappConfig',
+    'subscriptions.apps.SubscriptionsConfig',
+    'bootstrap5'
 ]
 
 MIDDLEWARE = [
@@ -140,3 +142,6 @@ DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
 LOGIN_REDIRECT_URL = "home"
 LOGOUT_REDIRECT_URL = "home" 
+
+EMAIL_BACKEND = "django.core.mail.backends.filebased.EmailBackend"
+EMAIL_FILE_PATH = BASE_DIR / "sent_emails"
