+++ recipeApp/urls.py
@@ -5,5 +5,5 @@ from . import views
 
 urlpatterns = [
     path("subscriberhome", views.subscriber_home, name="subscriber_home"),
-    path("shefhome", views.index, name="index"),
+    path("chefhome", views.index, name="index"),
 ]
